+++ src/components/parts/Branch.ts
@@ -1,4 +1,4 @@
-import { EmotionDataType } from './types/EmotionDataType';
+import { EmotionDataType } from '../../types/EmotionDataType';
 
 type emotionDataType = EmotionDataType;
 
@@ -8,7 +8,7 @@ const Branch = (emotionData: emotionDataType) => {
   let index = 0;
   for (const emotion in emotionData) {
     index++;
-    if (emotionData[emotion as keyof emotionDataType] > max && index < 5) {
+    if (emotionData[emotion as keyof emotionDataType] > max && index < 4) {
       max = emotionData[emotion as keyof emotionDataType];
       maxEmotion = index;
     }
