+++ src/components/parts/Eat.tsx
@@ -1,16 +1,16 @@
 import { FC } from 'react';
 import Box from '@mui/material/Box';
-import yoro_hana from './assets/yoro_hana.png';
-import yoro_kazitu from './assets/yoro_kazitu.png';
-import ikari_hana from './assets/ikari_hana.png';
-import ikari_kazitu from './assets/ikari_kazitu.png';
-import kana_hana from './assets/kana_hana.png';
-import kana_kazitu from './assets/kana_kazitu.png';
-import tano_hana from './assets/tano_hana.png';
-import tano_kazitu from './assets/tano_kazitu.png';
-import yagi_syokuzi from './assets/yagi_syokuzi.png';
-import yagi_left from './assets/yagi_left.png';
-import { EmotionDataType } from './types/EmotionDataType';
+import yoro_hana from '../../assets/yoro_hana.png';
+import yoro_kazitu from '../../assets/yoro_kazitu.png';
+import ikari_hana from '../../assets/ikari_hana.png';
+import ikari_kazitu from '../../assets/ikari_kazitu.png';
+import kana_hana from '../../assets/kana_hana.png';
+import kana_kazitu from '../../assets/kana_kazitu.png';
+import tano_hana from '../../assets/tano_hana.png';
+import tano_kazitu from '../../assets/tano_kazitu.png';
+import yagi_syokuzi from '../../assets/yagi_syokuzi.png';
+import yagi_left from '../../assets/yagi_left.png';
+import { EmotionDataType } from '../../types/EmotionDataType';
 
 type Props = {
   emotionData: EmotionDataType;
@@ -65,8 +65,7 @@ const Eat: FC<Props> = (props) => {
                   );
                 //怒
                 case 2:
-                  return random ===
-                   0 ? (
+                  return random === 0 ? (
                     <Box
                       sx={{
                         width: '70px',
