+++ src/components/parts/NavBar.tsx
@@ -4,7 +4,7 @@ import HelpIcon from '@mui/icons-material/Help';
 import SearchIcon from '@mui/icons-material/Search';
 import MusicNoteIcon from '@mui/icons-material/MusicNote';
 import MusicOffIcon from '@mui/icons-material/MusicOff';
-import bgm from './Audio/Bgm.mp3';
+import bgm from '../../Audio/Bgm.mp3';
 
 type Props = {
   handleTutorialChange: React.MouseEventHandler<HTMLButtonElement> | undefined;
