+++ src/routes/Routers.tsx
@@ -1,0 +1,12 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { AppView } from '../components/pages/AppView';
+
+export const Router = () => {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path={'/'} element={<AppView />}></Route>
+      </Routes>
+    </BrowserRouter>
+  );
+};
